add weather-information project
This project is a simple web application for displaying live weather information using the OpenWeatherMap API. It shows the weather data for multiple cities, including details like temperature, humidity, wind speed, and a description of the weather, all displayed in separate cards. The weather data is fetched in real-time from the API  with the Axios library and dynamically inserted into the user interface.